+++ mockingbird/src/main/kotlin/com/leftindust/mediq/dao/DoctorDao.kt
@@ -5,7 +5,7 @@ import com.leftindust.mediq.dao.entity.Doctor
 import com.leftindust.mediq.extensions.CustomResult
 
 interface DoctorDao {
-    suspend fun getByPatient(pid: Int, requester: MediqToken): CustomResult<List<Doctor>, OrmFailureReason>
+    suspend fun getByPatient(pid: Long, requester: MediqToken): CustomResult<List<Doctor>, OrmFailureReason>
     suspend fun getByVisit(vid: Long?, requester: MediqToken): CustomResult<Doctor, OrmFailureReason>
     suspend fun getByDoctor(did: Long, requester: MediqToken): CustomResult<Doctor, OrmFailureReason>
 }
