+++ mockingbird/src/main/kotlin/com/leftindust/mediq/dao/RecordDao.kt
@@ -6,5 +6,5 @@ import com.leftindust.mediq.extensions.CustomResult
 
 interface RecordDao {
     suspend fun getRecordByRecordId(rid: Int, requester: MediqToken): CustomResult<MediqRecord, OrmFailureReason>
-    suspend fun getRecordsByPatientPid(pid: Int, requester: MediqToken): CustomResult<List<MediqRecord>, OrmFailureReason>
+    suspend fun getRecordsByPatientPid(pid: Long, requester: MediqToken): CustomResult<List<MediqRecord>, OrmFailureReason>
 }
