+++ mockingbird/src/main/kotlin/com/leftindust/mediq/graphql/queries/VisitQuery.kt
@@ -5,6 +5,7 @@ import com.expediagroup.graphql.server.operations.Query
 import com.leftindust.mediq.auth.GraphQLAuthContext
 import com.leftindust.mediq.dao.VisitDao
 import com.leftindust.mediq.extensions.toInt
+import com.leftindust.mediq.extensions.toLong
 import com.leftindust.mediq.graphql.types.GraphQLVisit
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Controller
@@ -18,7 +19,7 @@ class VisitQuery(
         graphQLAuthContext: GraphQLAuthContext
     ): List<GraphQLVisit> {
         val requester = graphQLAuthContext.mediqAuthToken
-        return visitDao.getVisitsForPatientPid(pid.toInt(), requester)
+        return visitDao.getVisitsForPatientPid(pid.toLong(), requester)
             .getOrThrow()
             .map { GraphQLVisit(it, it.id!!, graphQLAuthContext) }
     }
